requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
PyPDF2
google-generativeai
pydantic
httpx